requirements.txt
PyPDF2>=3.0.1,<4.0.0
anthropic>=0.3.0,<0.4.0
python-dotenv>=0.19.2,<0.20.0
PyYAML>=6.0.1